requirements.txt
# Core LLM & Orchestration
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
httpx>=0.27.0
playwright>=1.44.0

# External Research
duckduckgo-search>=6.2.0

# Financial Data
curl-cffi>=0.7.0
pandas-datareader>=0.10.0
yfinance==0.2.53
websockets>=10.0,<13.0

# UI
gradio>=4.44.0,<5.0.0
huggingface_hub==0.25.2

# Utilities
python-dotenv>=1.0.0
tldextract>=5.1.0
